// src/lib/api/base/api.ts
export interface AllStrategyEnum {
}
// ...
export interface AppSchemasStrategyProfitAndLoss {
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'strategy_direction'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'max_profit'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'max_loss'?: any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'breakeven_points': any;
    /**
     * 
     * @type {any}
     * @memberof AppSchemasStrategyProfitAndLoss
     */
    'pop'?: any;
}
// ...
export interface PlotResult {
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'result'?: any;
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'spot_prices'?: any;
    /**
     * 
     * @type {any}
     * @memberof PlotResult
     */
    'strike_prices'?: any;
}
// ...
export interface StrategyDataInputBlueprint {
    /**
     * 
     * @type {any}
     * @memberof StrategyDataInputBlueprint
     */
    'name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyDataInputBlueprint
     */
    'tags'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyDataInputBlueprint
     */
    'data'?: any;
}
// ...
export interface StrategyDataUpdateInputBlueprint {
    /**
     * 
     * @type {any}
     * @memberof StrategyDataUpdateInputBlueprint
     */
    'id': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyDataUpdateInputBlueprint
     */
    'name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyDataUpdateInputBlueprint
     */
    'tags'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyDataUpdateInputBlueprint
     */
    'data'?: any;
}
/**
 * doc string 
 * @export
 * @interface StrategyEnum
 */
export interface StrategyEnum {
}
// ...
export interface StrategyPayOffData {
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_data'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price_entered'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'spot_price_diff'?: any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'index_name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'expiry_date': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'time': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_name': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'total_value': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_journey': any;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'strategy_exit_journey': any;
    /**
     * 
     * @type {AppSchemasStrategyProfitAndLoss}
     * @memberof StrategyPayOffData
     */
    'profit_and_loss': AppSchemasStrategyProfitAndLoss;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'script_lot_size': any;
    /**
     * 
     * @type {PlotResult}
     * @memberof StrategyPayOffData
     */
    'plot_results': PlotResult;
    /**
     * 
     * @type {any}
     * @memberof StrategyPayOffData
     */
    'pop'?: any;
}
// ...
export class OptionsStrategyBuilderApi extends BaseAPI {
    /**
     * 
     * @summary Create Strategy
     * @param {StrategyDataInputBlueprint} strategyDataInputBlueprint 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public createStrategyV1StrategiesPost(strategyDataInputBlueprint: StrategyDataInputBlueprint, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).createStrategyV1StrategiesPost(strategyDataInputBlueprint, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get Data For Custom Made Strategy
     * @param {any} strategyId 
     * @param {any} indexName 
     * @param {any} time 
     * @param {any} expiryDate 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public getDataForCustomMadeStrategyV1StrategiesByIdStrategyIdExampleGet(strategyId: any, indexName: any, time: any, expiryDate: any, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).getDataForCustomMadeStrategyV1StrategiesByIdStrategyIdExampleGet(strategyId, indexName, time, expiryDate, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get Data For Ready Made Strategy
     * @param {StrategyEnum} strategyName 
     * @param {any} indexName 
     * @param {any} time 
     * @param {any} expiryDate 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public getDataForReadyMadeStrategyV1StrategiesByNameStrategyNameExampleGet(strategyName: StrategyEnum, indexName: any, time: any, expiryDate: any, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).getDataForReadyMadeStrategyV1StrategiesByNameStrategyNameExampleGet(strategyName, indexName, time, expiryDate, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Get Ready Made Strategy By Name
     * @param {AllStrategyEnum} strategyName 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public getReadyMadeStrategyByNameV1StrategiesByNameStrategyNameGet(strategyName: AllStrategyEnum, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).getReadyMadeStrategyByNameV1StrategiesByNameStrategyNameGet(strategyName, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary List All Ready Made Strategies
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public listAllReadyMadeStrategiesV1StrategiesGet(options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).listAllReadyMadeStrategiesV1StrategiesGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Prepare And Get Simulator Data
     * @param {any} indexName 
     * @param {any} time 
     * @param {any} expiryDate 
     * @param {any} strategyName 
     * @param {any} body 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public prepareAndGetSimulatorDataV1StrategiesPayoffPost(indexName: any, time: any, expiryDate: any, strategyName: any, body: any, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).prepareAndGetSimulatorDataV1StrategiesPayoffPost(indexName, time, expiryDate, strategyName, body, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update Strategy
     * @param {StrategyDataUpdateInputBlueprint} strategyDataUpdateInputBlueprint 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof OptionsStrategyBuilderApi
     */
    public updateStrategyV1StrategiesPut(strategyDataUpdateInputBlueprint: StrategyDataUpdateInputBlueprint, options?: AxiosRequestConfig) {
        return OptionsStrategyBuilderApiFp(this.configuration).updateStrategyV1StrategiesPut(strategyDataUpdateInputBlueprint, options).then((request) => request(this.axios, this.basePath));
    }
}